OneDrive - Ashesi University/Documents/ASHESI/THIRD YEAR 2ND SEM/bitecheck-ml-model/requirements.txt
# Core dependencies
tensorflow>=2.9.0
numpy>=1.22.0
matplotlib>=3.5.0
Pillow>=9.0.0

# Data handling
pandas>=1.4.0
scikit-learn>=1.0.0

# Image processing
opencv-python>=4.5.0

# Model utilities
h5py>=3.6.0

# Development tools (optional)
ipykernel>=6.9.0
jupyter>=1.0.0